1-ai-agents/requirements.txt
azure-identity>=1.12.0
azure-mgmt-resource>=23.0.1
azure-mgmt-cognitiveservices>=13.5.0
